requirements.txt
Django
django-crispy-forms
django-filter
django-pure-pagination
django-chartjs
PyMySQL
mysqlclient
